Ex2.17
If you delete a person in browser 1 a short while before attempting to change the person's phone number in browser 2, you will get the following error message:

Fix the issue according to the example shown in promise and errors in part 2. Modify the example so that the user is shown a message when the operation does not succeed. The messages shown for successful and unsuccessful events should look different: