Ex 2.11 (new)
We continue with developing the phonebook. Store the initial state of the application in the file db.json, which should be placed in the root of the project.

Start json-server on port 3001 and make sure that the server returns the list of people by going to the address http://localhost:3001/persons in the browser.