Ex 3.1 (new)
Implement a Node application that returns a hardcoded list of phonebook entries from the address http://localhost:3001/api/persons.

Notice that the forward slash in the route api/persons is not a special character, and is just like any other character in the string.

The application must be started with the command npm start.

The application must also offer an npm run dev command that will run the application and restart the server whenever changes are made and saved to a file in the source code.