Ex 3.19-22
3.19 Expand the validation so that the name stored in the database has to be at least three characters long. Expand the frontend so that it displays some form of error message when a validation error occurs. Error handling can be implemented by adding a catch block

3.20 : Phonebook database, step8
Add validation to your phonebook application, which will make sure that phone numbers are of the correct form. A phone number must:

have length of 8 or more
be formed of two parts that are separated by -, the first part has two or three numbers and the second part also consists of numbers

eg. 09-1234556 and 040-22334455 are valid phone numbers
eg. 1234556, 1-22334455 and 10-22-334455 are invalid

3.21 Generate a new "full stack" version of the application by creating a new production build of the frontend, and copying it to the backend repository. Verify that everything works locally by using the entire application from the address http://localhost:3001/.

3.22 Add ESlint to your application and fix all the warnings.