Exer 6.12-13
6.12 Extend the component so that it renders the message stored in the Redux store, making the component take the following form
You will have to make changes to the application's existing reducer. Create a separate reducer for the new functionality by using the Redux Toolkit's createSlice function.

The application does not have to use the Notification component intelligently at this point in the exercises. It is enough for the application to display the initial value set for the message in the notificationReducer.

6.13 Extend the application so that it uses the Notification component to display a message for five seconds when the user votes for an anecdote or creates a new anecdote: