Exer 7.7
7.7 Use the code from https://github.com/fullstack-hy2020/country-hook as your starting point.
The application can be used to search for a country's details from the service in https://studies.cs.helsinki.fi/restcountries/. If a country is found, the details of the country are displayed. The application is otherwise complete, but in this exercise, you have to implement a custom hook useCountry, which can be used to search for the details of the country given to the hook as a parameter.
Use the API endpoint name to fetch a country's details in a useEffect hook within your custom hook.